requirements.txt
Flask
Flask-SocketIO
python-dotenv
google-generativeai
eventlet
